Moved flattening of types into UnionType class and removed logic from MappedTypeNodeParser.
